requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
psycopg[binary]
